\name{cat_function}
\alias{cat_function}
\title{A Cat Function}
\usage{
cat_function(love = TRUE)
}
\arguments{
  \item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
cat_function(love=FALSE)
}
\seealso{
Other fastpackage: \code{\link{fastLSA}};
\code{\link{fastpackage}},
\code{\link{fastpackage-package}}; \code{\link{mutlinksC}};
\code{\link{subdiagC}}
}
\keyword{cats}

